api/requirements.txt
Django==4.2.11
django-cors-headers==4.4.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-filter==24.2
drf-spectacular==0.27.2
pytest==8.2.2
pytest-django==4.8.0
factory-boy==3.3.0
